move `JavaClassProcessor.DeclarationHandler` toplevel
This is meanwhile also used by `JavaClassSignatureImporter`, so it is a strange location to have it declared within `JavaClassProcessor` which in turn uses the `JavaClassSignatureImporter`. This creates an unnecessary cyclic dependency between the files.
